+++ src/impls/other.rs
@@ -37,7 +37,6 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         _initial_location: Option<&FileUri>,
         _mime_types: &[&str],
         _multiple: bool,
-        _take_persistable_uri_permission: bool
     ) -> crate::Result<Vec<FileUri>> {
 		
         Err(crate::Error::NotAndroid)
@@ -47,7 +46,6 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         &self,
         _target: VisualMediaTarget,
         _multiple: bool,
-        _take_persistable_uri_permission: bool
     ) -> crate::Result<Vec<FileUri>> {
 		
         Err(crate::Error::NotAndroid)
@@ -58,7 +56,6 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         _initial_location: Option<&FileUri>,
         _initial_file_name: impl AsRef<str>,
         _mime_type: Option<&str>,
-        _take_persistable_uri_permission: bool
     ) -> crate::Result<Option<FileUri>> {
 
         Err(crate::Error::NotAndroid)
@@ -101,13 +98,28 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     fn show_manage_dir_dialog(
         &self,
         _initial_location: Option<&FileUri>,
-        _take_persistable_uri_permission: bool
     ) -> crate::Result<Option<FileUri>> {
         Err(crate::Error::NotAndroid)
     }
 
     fn app_handle(&self) -> &tauri::AppHandle<R> {
         &self.0
+    }
+    
+    fn take_persistable_uri_permission(&self, _uri: &FileUri, _mode: PersistableAccessMode) -> crate::Result<()> {
+        Err(crate::Error::NotAndroid)
+    }
+    
+    fn get_all_persisted_uri_permissions(&self) -> crate::Result<impl Iterator<Item = PersistedUriPermission>> {
+        Err::<std::iter::Empty<_>, _>(crate::Error::NotAndroid)
+    }
+    
+    fn release_persisted_uri_permission(&self, _uri: &FileUri) -> crate::Result<()> {
+        Err(crate::Error::NotAndroid)
+    }
+    
+    fn release_all_persisted_uri_permissions(&self) -> crate::Result<()> {
+        Err(crate::Error::NotAndroid)
     }
 }